remove the (double) cast, changed all the penalty defend pos to double

+++ module/purpose/Goalie/src/Goalie.hpp
@@ -43,9 +43,9 @@ namespace module::purpose {
         /// @brief Stores configuration values
         struct Config {
             /// @brief Ready position to walk to (x, y, theta)
-            Eigen::Vector3f ready_position = Eigen::Vector3f::Zero();
+            Eigen::Vector3d ready_position = Eigen::Vector3f::Zero();
             /// @brief Penalty defence position to walk to when defending from a penalty kick from the opposing team
-            Eigen::Vector3f penalty_defence_position = Eigen::Vector3f::Zero();
+            Eigen::Vector3d penalty_defence_position = Eigen::Vector3f::Zero();
             /// @brief Ready position to walk to
             Eigen::Isometry3d Hfr = Eigen::Isometry3d::Identity();
             /// @brief x minimum bound on field to walk within
